Extract tools and utils to dedicated files
