+++ packages/chookscord/src/tools/changed/index.ts
@@ -1,1 +1,2 @@
+export * from './choices';
 export * from './options';
