--- packages/chookscord/src/utils/is-same/_object.ts
@@ -1,12 +1,0 @@
-export function keysAreSame<T>(
-  a: T,
-  b: T,
-  keys: (keyof T)[],
-): boolean {
-  for (const key of keys) {
-    if (a[key] !== b[key]) {
-      return false;
-    }
-  }
-  return true;
-}
